mm: allow disabling on-demand memory strategies
This commit introduces an API to disable kernel mechanisms/strategies
which tend to desync the virtual memory view from the page table state
for a particular memory context. Typically, copy-on-write and
on-demand PTE updates are among these mechanisms.

This feature is useful when the delay incurred in handling a processor
fault for fixing up the current memory context is not acceptable for
some use cases (e.g. real-time configurations).

CAUTION: enabling this option for a process induces a significant
overhead for the fork(2) system call when used by any thread which
belongs to such context.