ARM/mm: allow pinning PTEs for kernel mappings
Some profiles (e.g. real-time) on armv4,v5 cannot cope with the
overhead induced by minor faults when treading on kernel memory.

When CONFIG_FORCE_COMMIT_MEMORY is enabled, PTEs referring to kernel
mappings are pinned in the page table.

NOTE: this mechanism is not enabled for armv6 and above, for which
handling a minor fault is reasonably cheap.