mm: irq_pipeline: lock out oob stage while updating PTE range
In the non-numa case, change_pte_range() will mark each updated PTE in
a range as absent, while the protbits are being fixed up, to make sure
it won't race with the hardware.

We don't want any oob code to preempt the in-band context until the
updated protbits are written back, so we have to hard disable IRQs
around this section.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>