dovetail: add core support
This is the basic support code for connecting the host Linux kernel
with a high-priority, external co-kernel as a guest, aka
"dovetailing".

The co-kernel leverages the interrupt pipeline ability to turn common
device IRQs into pseudo-NMIs from the perspective of the host kernel
(CONFIG_IRQ_PIPELINE), to achieve very low latency.