locking: irq_pipeline: add mutable locking API
A mutable spinlock combines real and virtual IRQ disabling when
protecting a critical section over the root stage, disabling
preemption like regular raw spinlocks. It behaves like a hard spinlock
assuming that hard IRQ disabling is already in effect when used from
the pipeline entry context, which may preempt any stage.

The idea behind this is to automatically convert raw spinlocks to hard
ones for short locked sections, optimizing the fast path when
injecting external IRQs to the pipeline.

We typically use this with the irqdesc lock.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>