irq_pipeline: add proxy tick device
The proxy tick device is a synthetic clock event device for handing
over control of the real/physical tick device in use by the host
kernel to a co-kernel.

This allows the co-kernel to control the real device, for handling
timing requests from the host kernel in addition to its own timing
duties.