ARM: steely: remove doxygen-based Cobalt documentation
The kernel documentation is moving to Sphinx which uses a reST
markup. In addition, much of the former Cobalt documentation won't
apply to the steely core in the end, as both are likely to diverge in
many ways amid the effort to better integrate the later into the
mainline kernel.