steely: keep usage of clocksource frequency strictly internal
Now that userland stopped dealing with raw ticks (aka TSC), we can
stop exposing the clocksource frequency through the feature block
during the process->core binding operation.

At this chance, we also get rid of the ancient .clockfreq kernel
parameter since the kernel always knows better about the actual
frequency, which the user should not be allowed to mess with.

Support for dealing with clocksource frequencies is scheduled for
removal when the core uses nanoseconds internally instead of clock
ticks.