steely/irq: drop the xnintr abstraction
We don't need or want this indirection, the interrupt management logic
exclusively belongs to the regular genirq layer. The per-IRQ stats
should be reintroduced later directly into genirq for pipelined IRQs.

Since genirq supports shared pipelined IRQs, there is no point in
providing any special handling for those in the steely core.

The former "edge IRQ" support in Cobalt used to iterate over the daisy
chain of handlers until a complete acknowledge round returned without
handling the IRQ. The purpose of this was unclear, and this definitely
requires further investigation if this was about fixing a potential
loss of edge events (e.g. condition raised on the device without IRQ
notification).