steely/procfs: drop "faults" vfile
We don't need to maintain a complex infrastructure for tracking and
reporting faults taken in primary mode, that is completely overkill:
since they are always pathological, we just need to report them in the
kernel log, since this sufficient to deliver the hint that something
is going wrong to the user.

At this chance, simplify the dovetail trap handler in the same move.