geninrq: irq_pipeline: attempt to reschedule on root enable
If CONFIG_PREEMPT is enabled, try to reschedule after we have played
the logged IRQ, so that we don't miss a scheduling opportunity.

This is required because we won't always go through the slow path
returning to userland when unwinding the context, checking for
TIF_WORK (and thus need_resched()). Typically, a synthetic interrupt
logged by a driver and deferred (due to root stall) while running
kernel code would not follow such path.

Likewise, a real interrupt preempting a stalled section of the regular
kernel would not traverse the rescheduling point in the assembly code,
but rather take the fast IRQ exit path.