ARM: dovetail: enable VDSO access from the oob stage
Co-kernel threads running in user-space may want to access the VDSO
page from the head stage, e.g. to call clock-related fast services. In
such a case, relying on the sequence count exclusively to order R/W
accesses is unsafe, in case update_vsyscall() is preempted by such
thread on the same CPU via an OOB interrupt.

i.e. there would be no way for the update sequence running over the
in-band stage to finish, causing the thread to loop indefinitely over
the oob stage.

To make it safe, update_vsyscall() must disable hard irqs while
updating the VDSO page, in order to exclude any preemption. The
sequence count is still relevant to inform remote CPUs that an update
is in progress though.