dovetail: stop caring about the modunload vs hook race
Dovetail should not be running a hook located in the text segment of a
module which is being unloaded.  This race must now be dealt with by
any modular co-kernel, by disabling all reasons for Dovetail to fire a
hook then synchronizing, prior to returning from the exit routine of
the container module.