dovetail: preserve hard interrupt state across mayday trap
The mayday hook asks the co-kernel to deal with a runaway context
which has been detected earlier. When doing so, the co-kernel might
re-enable hard interrupts, which we have to disable again before
unwinding the IRQ context which triggered the whole process (like a
watchdog event).

Go through dovetail_call_mayday() to preserve the original interrupt
state instead of open coding the call to the mayday hook. In addition,
this makes sure that the latter is always called with hard irqs off.