evl/monitor: add ungated signal-wait operations
This is a special form of the wait operation which is not protected by
a lock. Userland is assumed to deal with signal-vs-wait races in its
own way.

An event value is now available from the monitor's state, which can be
tested for signedness as the fixed condition for waiting, which comes
in handy for implementing semaphore behavior. Plain flag-type behavior
could be implemented the same way.