time: proxy: simplify the user interface for setup
There is no reason to ask the user for allocating the proxy device, we
just need it to finalize the prep work before the proxy gets
registered in the clock framework.

Turn get_percpu_device() into setup_proxy(), which also allows us to
pre-init the proxy descriptor with the real device settings before
calling the user setup handler instead of running post-init code
afterwards. This definitely makes the picture much clearer on the
user's end about which proxy settings will be used eventually.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>