evl/thread: handle EVL_THRIOC_SWITCH_OOB from OOB context only
The only sane way to issue a request for switching OOB via
evl_switch_oob() is through oob_ioctl(), not ioctl(). Once the switch
is done, the .unlocked_ioctl() operation would unwind into the VFS on
the out-of-band stage, which would certainly be a bad idea.

To reflect this, process EVL_THRIOC_SWITCH_OOB from oob_ioctl()
exclusively.