x86: dovetail: keep updates to TLS register fully atomic
We must not allow an out-of-band task to preempt an inband task while
the latter updates %fs / %gs, otherwise the in-register (MSR) value
might not match the in-memory copy, which is kind of problematic when
switching back later to the inband context.

To fix this, hard disable preemption around ARCH_SET_FS and
ARCH_SET_GS operations.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>