evl/monitor: plan for extending the range of supported protocols
Introduce a 'protocol' specifier for monitors along with a basic type
(gate or event), which userland can pass to change the way a monitor
can be signaled and awaited. Define PI/PP gates using this new
information bit associated to the generic EVL_MONITOR_GATE type.

With this change in, the pre-existing ungated event can be enriched
with additional protocols beyond the semaphore-based P/V operations.