genirq: irq_pipeline: synchronize pipeline on last pending IRQ
We currently synchronize the pipeline unconditionally after each
iteration of the interrupt decoding loop by the irqchip driver, which
almost always involves playing in-band IRQs whenever some events are
logged.

Instead of this, give a bonus to external out-of-band IRQs which are
immediately handled on pipeline entry but might be pending at the same
time than in-band events in the irqchip, by saving the time required
to run the synchronization code until the last IRQ is pulled out from
the interrupt controller.

Likewise, the companion kernel should be notified of the interrupt
frame unwinding via exit_oob_irq() only when the last IRQ has been
decoded, so that out-of-band rescheduling can happen only after the
interrupt level is back to normal in the irqchip.

Define enter_irq_pipeline() and leave_irq_pipeline() as generic
services for performing the necessary actions.