genirq: irq_pipeline: check_inband() must preserve the interrupt state
If check_inband() trips, both the in-band and oob interrupt states
must be preserved, not to add even more problems to the
picture. Specifically, we should not call irq_pipeline_oops() which
forcibly disables interrupts virtually and hardware-wise in a panic
path.

check_inband() is not - yet - a panic path, and as such should not
disable in-band interrupts when the oob stage is found stalled when it
should not be. However, we still want to raise the oopsing flag in
order to disable further interrupt-related checks.

This issue had weird consequences like this one:

hard_lock_irq(&lock);
  /* !irqs_disabled(), hard_irqs_disabled(), oob stalled */
  spin_lock_irqsave(&lock2, flags);
     flags = inband_irqsave()
         check_inband() -> local_irq_disable() via irq_pipeline_oops()
         flags eventually denotes a stalled in-band stage
  spin_lock_irqrestore(&lock2, flags);
  /* irqs_disabled() */

At this chance, clarify the related warning message too.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>