dovetail: altsched: fix preemption count breakage
Tasks running out-of-band may alter the (in-band) preemption count as
long as they don't trigger an in-band rescheduling, which Dovetail
properly blocks.

If the preemption count is not stack-based but a global per-cpu
variable instead, changing it has a globally visible side-effect
though, which is a problem if the out-of-band task is preempted and
schedules away before the change is rolled back: this may cause the
in-band context to later resume with a broken preemption count.

For this reason, the preemption count of any context which blocks from
the out-of-band stage is carried over and restored across switches,
emulating a stack-based storage.

Eventually, the count is reset to FORK_PREEMPT_COUNT upon transition
from out-of-band to in-band stage, reinstating the value in effect
when the converse transition happened at some point before.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>