genirq: irq_pipeline: fix handling of edge events
IRQ events which occur on a signal edge may be latched again as soon
as the current event is acked in the controller, which means while the
latter is being processed in do_oob_irq(). We need to iterate over
do_oob_irq() until all edge events pending for a given IRQ line
received by any CPU have been handled.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>