sched: dovetail: detect inband->oob transition epilogue
During an inband->oob transition, and the premption count is a
stack-based information, running_inband() and running_oob() become
out-of-sync at context switch point until the stage offset is updated
in inband_switch_tail(), which means we cannot use them in the latter
to figure out which stage we are currently running on.

To solve this, test _TLF_OFFSTAGE in order to determine whether we are
coming from dovetail_context_switch() for an inband->oob transition
(set), or from the regular context_switch() otherwise.